requirements.txt
Flask
gunicorn
openai
tiktoken
hdbcli
Pillow
opencv-python-headless
pytesseract 
python-dotenv